+++ libs/tokens/src/state/tokens/allTokensAtom.ts
@@ -11,7 +11,7 @@ import { lowerCaseTokensMap } from '../../utils/lowerCaseTokensMap'
 import { parseTokenInfo } from '../../utils/parseTokenInfo'
 import { tokenMapToListWithLogo } from '../../utils/tokenMapToListWithLogo'
 import { environmentAtom } from '../environmentAtom'
-import { listsEnabledStateAtom, listsStatesListAtom } from '../tokenLists/tokenListsStateAtom'
+import { listsEnabledStateAtom, listsStatesListAtom, tokenListsUpdatingAtom } from '../tokenLists/tokenListsStateAtom'
 
 export interface TokensByAddress {
   [address: string]: TokenWithLogo | undefined
@@ -78,6 +78,7 @@ export const allActiveTokensAtom = atom(async (get) => {
   const { chainId, enableLpTokensByDefault } = get(environmentAtom)
   const userAddedTokens = get(userAddedTokensAtom)
   const favoriteTokensState = get(favoriteTokensAtom)
+  const isTokenListsUpdating = get(tokenListsUpdatingAtom)
 
   const { tokensState: tokensMap, listsCount } = await get(tokensStateAtom)
   const nativeToken = NATIVE_CURRENCIES[chainId]
@@ -85,7 +86,7 @@ export const allActiveTokensAtom = atom(async (get) => {
   /**
    * Wait till token lists loaded
    */
-  if (listsCount === 0) {
+  if (!isTokenListsUpdating ? false : listsCount === 0) {
     return { tokens: [], chainId }
   }