+++ libs/hook-dapp-lib/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.11.0](https://github.com/cowprotocol/cowswap/compare/hook-dapp-lib-v1.10.0...hook-dapp-lib-v1.11.0) (2025-08-28)
+
+
+### Features
+
+* **trade:** implement partial approve ([#6139](https://github.com/cowprotocol/cowswap/issues/6139)) ([6f9d6b0](https://github.com/cowprotocol/cowswap/commit/6f9d6b055323d25705ee88a533f71190944be6e2))
+
+
+### Bug Fixes
+
+* **lens:** follow up fixes ([#6176](https://github.com/cowprotocol/cowswap/issues/6176)) ([cd8f3c0](https://github.com/cowprotocol/cowswap/commit/cd8f3c0201b8f04fc31a59588b6d2d05d215e112))
+
 ## [1.10.0](https://github.com/cowprotocol/cowswap/compare/hook-dapp-lib-v1.9.0...hook-dapp-lib-v1.10.0) (2025-08-12)